+++ R/print.R
@@ -19,9 +19,9 @@ print.dbreg = function(x, fes = FALSE, ...) {
     if (x$strategy == "compress") {
       cat("Compressed OLS estimation, Dep. Var.:", x$yvar, "\n")
       cat("Observations.:", prettyNum(x$nobs_orig, big.mark = ","), "(original) |", prettyNum(x$nobs, big.mark = ","), "(compressed)", "\n")
-    } else if (x$strategy == "mundlak") {
+    } else if (x$strategy == "demean") {
       num_fes = length(x$fes)
-      mstring = "Mundlak"
+      mstring = "Within"
       if (num_fes == 1) {
         mstring = paste("One-way", mstring)
       } else if (num_fes == 2) {